+++ scripts/themes/pullThemesFromMarketplace.ts
@@ -4,7 +4,7 @@ import { downloadFromMarketplace } from '../util/downloadFromMarketplace'
 go()
 
 async function go() {
-  for (let [extFullId, fPaths] of Object.entries(marketplaceThemeSources)) {
-    await downloadFromMarketplace(extFullId, fPaths, 'theme')
+  for (const [extFullId, themes] of Object.entries(marketplaceThemeSources)) {
+    await downloadFromMarketplace(extFullId, themes, 'theme')
   }
 }